+++ Sources/ParseSwift/API/URLSession+extensions.swift
@@ -66,6 +66,11 @@ extension URLSession {
                           let json = try? JSONSerialization
                             .data(withJSONObject: responseData,
                               options: .prettyPrinted) else {
+                        let nsError = error as NSError
+                        if nsError.code == 4865,
+                          let description = nsError.userInfo["NSDebugDescription"] {
+                            return .failure(ParseError(code: .unknownError, message: "Invalid struct: \(description)"))
+                        }
                         return .failure(ParseError(code: .unknownError,
                                                    // swiftlint:disable:next line_length
                                                    message: "Error decoding parse-server response: \(response) with error: \(error.localizedDescription) Format: \(String(describing: String(data: responseData, encoding: .utf8)))"))